# Conflicts:
#	src/main/kotlin/environment/vfs/Vfs.kt

+++ src/main/kotlin/helpers/ConfigStructure.kt
@@ -4,7 +4,12 @@ package helpers
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class ConfigStructure(val stackSize: Int, val memorySize: Int, val hertz: Int, val locations: List<Location>)
+data class ConfigStructure(
+	val stackSize: Int,
+	val memorySize: Long,
+	val hertz: Int,
+	val locations: List<Location>,
+)
 
 @Serializable
 data class Location(val requirement: String, val path: String)
